# Backend/jiufang-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 九方智能查询系统 API
  description: |
    九方智能查询系统是一个基于AI Agent的自然语言查询系统，支持用户通过自然语言查询ERP数据，
    并提供定时报告、预警规则、数据导出等功能。
    
    ## 认证方式
    - 使用JWT Bearer Token认证
    - 登录接口获取JWT令牌
    - 令牌有效期24小时
    
    ## 权限控制
    - admin: 系统管理员，拥有所有权限
    - manager: 管理层，拥有查询和导出权限
    - executive: 执行层，拥有查询权限
    
    ## 响应格式
    所有接口统一使用以下响应格式：
    ```json
    {
      "code": 200,
      "message": "success",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: 九方技术团队
    email: tech@jiufang.com
servers:
  - url: https://api.jiufang.com/v1
    description: 生产环境
  - url: https://api-dev.jiufang.com/v1
    description: 开发环境
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
paths:
...
  /query:
    post:
      tags:
        - Query
      summary: 自然语言查询
      description: 执行自然语言查询，支持对话上下文和权限过滤
      operationId: executeQuery
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - input
              properties:
                input:
                  type: string
                  minLength: 1
                  maxLength: 500
                session_id:
                  type: string
                  pattern: '^\d{19}$'
                execute_immediately:
                  type: boolean
                  default: true
            example:
              input: "查询上个月的销售额"
              session_id: "1234567890123456789"
              execute_immediately: true
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== 用户组管理模块 ====================
...
  /reports:
    post:
      tags:
        - Report
      summary: 创建定时报告
      description: 创建定时报告（仅系统管理员可操作）
      operationId: createReport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - query
                - schedule
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                query:
                  type: string
                schedule:
                  type: string
                  description: Cron表达式
                recipients:
                  type: array
                  items:
                    type: string
                    format: email
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags:
        - Report
      summary: 查询定时报告列表
      description: 分页查询定时报告列表（仅系统管理员可操作）
      operationId: listReports
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /reports/{id}:
    get:
      tags:
        - Report
      summary: 查询定时报告详情
      description: 查询定时报告详情（仅系统管理员可操作）
      operationId: getReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Report
      summary: 更新定时报告
      description: 更新定时报告（仅系统管理员可操作）
      operationId: updateReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                query:
                  type: string
                schedule:
                  type: string
                recipients:
                  type: array
                  items:
                    type: string
                    format: email
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Report
      summary: 删除定时报告
      description: 删除定时报告（仅系统管理员可操作）
      operationId: deleteReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /push-records:
    get:
      tags:
        - Report
      summary: 查询推送记录
      description: 查询推送记录列表（仅系统管理员可操作）
      operationId: listPushRecords
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushRecordListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== 预警规则模块 ====================
...
components:
  schemas:
    SuccessResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 错误码
          example: 40001
        message:
          type: string
          description: 错误信息
          example: "参数错误"
...
    QueryResultResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "查询成功"
        data:
          type: object
          required:
            - session_id
            - understanding
            - result_type
            - can_export
          properties:
            session_id:
              type: string
              pattern: '^\d{19}$'
            understanding:
              type: string
            result_type:
              type: string
              enum: [table, chart, empty]
            columns:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  type:
                    type: string
                    enum: [string, number, date]
            rows:
              type: array
              items:
                type: object
            chart_config:
              type: object
            suggested_questions:
              type: array
              items:
                type: string
            can_export:
              type: boolean
...
    ReportResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - id
            - name
            - query
            - schedule
            - created_at
          properties:
            id:
              type: string
              pattern: '^\d{19}$'
            name:
              type: string
            query:
              type: string
            schedule:
              type: string
            recipients:
              type: array
              items:
                type: string
                format: email
            created_at:
              type: string
              format: date-time
    ReportListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - reports
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            reports:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  name:
                    type: string
                  schedule:
                    type: string
                  created_at:
                    type: string
                    format: date-time
    PushRecordListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - records
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            records:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  report_id:
                    type: string
                    pattern: '^\d{19}$'
                  status:
                    type: string
                    enum: [success, failed]
                  created_at:
                    type: string
                    format: date-time
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40001
            message: "参数错误"
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40101
            message: "未授权访问"
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40301
            message: "权限不足"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40401
            message: "资源不存在"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证